packages/source_ingestion_tools/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "source-ingestion-tools"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "sqlalchemy[asyncio]>=2.0.36",
  "asyncpg>=0.30.0",
  "psycopg[binary]>=3.2.0",
  "google-generativeai>=0.8.3",
  "tenacity>=9.0.0",
  "pydantic-settings>=2.6.0",
  "pgvector>=0.3.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["source_ingestion_tools*"]
